Move PatternResources into NodaTime.Globalization, and AssemblyInfo.cs into the root
Goodbye Properties directory!
