+++ src/NodaTime/Globalization/PatternResources.resx
